refactor: views/Consent.vue -> views/Register.vue
change name of Consent.vue -> Register.vue to match /register path
