+++ client/src/router.ts
@@ -5,7 +5,7 @@ import Lobby from '@port-of-mars/client/views/Lobby.vue';
 import Game from '@port-of-mars/client/views/Game.vue';
 import SignedUp from '@port-of-mars/client/views/SignedUp.vue';
 import Tutorial from '@port-of-mars/client/views/Tutorial.vue';
-import Consent from '@port-of-mars/client/views/Consent.vue';
+import Register from '@port-of-mars/client/views/Register.vue';
 import PlayerDashboard from '@port-of-mars/client/views/Dashboard.vue';
 import Verify from '@port-of-mars/client/views/VerifyEmail.vue';
 import {
@@ -14,7 +14,7 @@ import {
   LOBBY_PAGE,
   GAME_PAGE,
   TUTORIAL_PAGE,
-  CONSENT_PAGE,
+  REGISTER_PAGE,
   DASHBOARD_PAGE,
   VERIFY_PAGE, SIGNEDUP_PAGE,
 } from '@port-of-mars/shared/routes';
@@ -29,7 +29,7 @@ const router = new Router({
     { ...PAGE_META[GAME_PAGE], component: Game },
     { ...PAGE_META[SIGNEDUP_PAGE], component: SignedUp },
     { ...PAGE_META[TUTORIAL_PAGE], component: Tutorial },
-    { ...PAGE_META[CONSENT_PAGE], component: Consent },
+    { ...PAGE_META[REGISTER_PAGE], component: Register },
     { ...PAGE_META[DASHBOARD_PAGE], component: PlayerDashboard },
     { ...PAGE_META[VERIFY_PAGE], component: Verify },
   ],
